backend/requirements.txt
fastapi[standard]
pydantic
chromadb
python-dotenv
groq
sentence-transformers
pdfplumber
edge_tts
python-multipart
openai
